// src/model/CommentWithActions.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Action } from "./Action";
import type { Comment } from "./Comment";

export type CommentWithActions = { comment: Comment, actions: Array<Action>, seen: boolean, submitted: boolean, };